moved calculation into EstimatedItemValueCalculator
