+++ src/main/java/at/hannibal2/skyhanni/features/misc/items/EstimatedWardrobePrice.kt
@@ -37,7 +37,7 @@ class EstimatedWardrobePrice {
         var totalPrice = 0.0
         for (item in items) {
             val name = item.name
-            val price = EstimatedItemValue.getEstimatedItemPrice(item, mutableListOf()).first
+            val price = EstimatedItemValueCalculator.calculate(item, mutableListOf()).first
             totalPrice += price
 
             toolTip.add(index++, "  §7- $name: §6${NumberUtil.format(price)}")
@@ -55,7 +55,7 @@ class EstimatedWardrobePrice {
 
         for ((slot, item) in event.inventoryItems) {
             item.getInternalNameOrNull() ?: continue
-            val price = EstimatedItemValue.getEstimatedItemPrice(item, mutableListOf()).first
+            val price = EstimatedItemValueCalculator.calculate(item, mutableListOf()).first
             if (price == 0.0) continue
             val id = slot % 9
             val list = map.getOrPut(id) { mutableListOf() }
